chore: move error into fred-types
