+++ fred-types/Cargo.toml
@@ -6,4 +6,10 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-redis-protocol = "0.1.2"
+redis-protocol = { version = "0.1.2", optional = true }
+url = { version = "1.7.2", optional = true }
+futures = { version = "0.1", optional = true }
+tokio-timer-patched = { version = "0.1.3", optional = true }
+
+[features]
+legacy = [ "dep:futures", "dep:tokio-timer-patched", "dep:redis-protocol", "dep:url" ]
