+++ fred-types/src/utils.rs
@@ -1,4 +1,4 @@
-use crate::types::RedisTypeError;
+use crate::error::{RedisError, RedisErrorKind};
 
 pub(crate) fn incr_with_max(curr: u32, max: u32) -> Option<u32> {
     if max == 0 {
@@ -11,12 +11,14 @@ pub(crate) fn incr_with_max(curr: u32, max: u32) -> Option<u32> {
 }
 
 /// Convert a redis string to an `f64`, supporting "+inf" and "-inf".
-pub(crate) fn redis_string_to_f64(s: &str) -> Result<f64, RedisTypeError> {
+pub(crate) fn redis_string_to_f64(s: &str) -> Result<f64, RedisError> {
     if s == "+inf" {
         Ok(f64::INFINITY)
     }else if s == "-inf" {
         Ok(f64::NEG_INFINITY)
     }else{
-        s.parse::<f64>().map_err(|_| RedisTypeError(format!("Could not convert {} to floating point value.", s)))
+        s.parse::<f64>().map_err(|_| RedisError::new(
+            RedisErrorKind::Unknown, format!("Could not convert {} to floating point value.", s)
+        ))
     }
 }
